// src/types.ts
export interface ResultList {
  alternative: Array<string>
  bookingId: string
  city: string
  country: string
  countryIso: string
  iata: string
  index: number
  isPopular: boolean
  lang: string
  lat: number
  lng: number
  locationId: string
  name: string
  placeKey: string
  placeType: string
  region: string
  searchType: string
  ufi: number
}